remove u128 from_lt

+++ packages/nekoton_bridge/rust/src/nekoton_wrapper/transport/mod.rs
@@ -214,8 +214,8 @@ impl TransportBoxTrait for name {
     ) -> anyhow::Result<String> {
         let address = parse_address(address)?;
         let from_lt = match from_lt {
-            None => u128::MAX,
-            Some(amount) => amount.parse::<u128>().handle_error()?,
+            None => u64::MAX,
+            Some(amount) => amount.parse::<u64>().handle_error()?,
         };
 
         let raw_transactions = self
@@ -467,8 +467,8 @@ impl TransportBoxTrait for GqlTransportBox {
         let address = parse_address(address)?;
 
         let from_lt = match from_lt {
-            None => u128::MAX,
-            Some(amount) => amount.parse::<u128>().handle_error()?,
+            None => u64::MAX,
+            Some(amount) => amount.parse::<u64>().handle_error()?,
         };
 
         let raw_transactions = self
